+++ oshi-core/src/test/java/oshi/util/platform/windows/PerfDataUtilTest.java
@@ -6,15 +6,16 @@ package oshi.util.platform.windows;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
-import static org.slf4j.event.Level.ERROR;
-import static org.slf4j.event.Level.WARN;
+import static oshi.util.LogLevel.ERROR;
+import static oshi.util.LogLevel.WARN;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.EnabledOnOs;
 import org.junit.jupiter.api.condition.OS;
-import org.slf4j.event.Level;
 
 import com.sun.jna.platform.win32.PdhMsg;
+
+import oshi.util.LogLevel;
 
 @EnabledOnOs(OS.WINDOWS)
 class PerfDataUtilTest {
@@ -39,7 +40,7 @@ class PerfDataUtilTest {
 
     @Test
     void testHandleErrorAtEveryLevel() {
-        for (Level level : Level.values()) {
+        for (LogLevel level : LogLevel.values()) {
             assertThat(PerfDataUtil.handleError(PdhMsg.PDH_INVALID_HANDLE, level, "Failed.", "default"), is("default"));
         }
     }